requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
tensorflow
numpy==1.26.4
Pillow==10.3.0
streamlit==1.35.0
requests==2.32.3
matplotlib==3.8.4
reportlab==4.2.0
